+++ formula.h
@@ -5,7 +5,7 @@
 #define append_operator(y) append_formula(0,y)
 #define append_value(y)    append_formula(1,y)
 
-int append_formula(int, int);
-int get_formula(int *);
+int append_formula(int, double);
+int get_formula(double *);
 
 #endif // __FORMULA_H__
